Tiled Map added
Hey guys added the tiled map. I will be adding a controller today so all
the instances can communicate with each other. feel free to work on your
parts, pathfinding, etc. If someone wants to work on a UML or diagram
that will be good to get everyone on the same page and we can assign
each other tasks. =)