Added Character scripts, see description
I added the character scripts, still some to do on it but a lot is done
in the basic sense. The resource script is extremely basic and used by
the character script. It can be expanded upon later if needed.
I moved some of the prefabs to a Resources folder to take advantage of
loading them and using them for references for instantiating prefabs in
the code.
Finally, I created a scene called "CharacterTestScene" that has a test
file for testing these scripts and examples on how they might be used.
Just open the Test.cs file and look through the Update function for the
keys and what they do.