Work on the Ally system.
The ally list is similar but different. It's its own script class now so
it's easy to get an object by it's index and get its index by the
object. The system no longer destroys the ally objects. This would be
the job of a controller to get the objects tagged as allies and destroy
them and not the characters. There are 3 more in-depth tests and I also
fixed the prefabs. They had sprites attached instead of game objects.