MapEvent 80% Done Item Class + Support Added
Finished most of the MapEvent. Missing are
-Tile triggers, still using N and R
-Resource interaction (missing fixed resources)
-Ally and Item support (errors/TODOs listed in appropriate areas)
-Enemy does not appear on screen (damn graphics)
-Weather needs access to map variable (can discuss in our meeting)
-Maybe skew probability of which event occurs (Make NOTHING more common
than ITEM, etc)

Item Class created. Support for items added to Character Class, but
needs more work and customization (TODOs listed in character class).
Predefined items should be expended in the future, but we'll keep it
simple for now.
Cost variable is there for when we add shops.

Also added inventory for character, meaning that resources will be
accounted by number and weight (which is how the resource management
comes into play). Since I added a size parameter, I updated the
associated test function. Currently everything runs but there are
warnings which might be worth looking into.