State Machine Skeleton redone + Fixed Resource Err
After my conversation with Javier and the agreed upon UMLs, I
restructured the state class so it followed the design indicated by the
UMLs. However, I am unable to see whether it works because the script
cannot be found when I run it in Unity. I'm not sure how to fix this but
the code is good. There was a conflict error in the Resource class so I
corrected that. I will resume coding the state machine tomorrow,
hopefully the script error can be resolved so testing can start.