Replaced GameplayStateMachine with StateMachine
Also created basic flow for game driven through key presses at the
moment. While you're playing it, note that the GUI Text in the game is
the current state of the game, whereas the console text is one cycle
behind.  (This will be fixed when the wait is removed, which is only
there for testing and visualization purposes).
Also had to update the area because the GameplayStateMachine was not
being read on my end, so I had to replace it with a new file,
StateMachine. If this will cause a conflict just make a new file on your
end and copy my code into it.